Transition to accounts and groups based wallet
Until now this was a currencies centric wallet and we could
add only a single account for each currency.

Now the whole thing is rewritten so that we can have as many
accounts as we want in any given currency and we have those
accounts grouped in some predefined groups.